requirements/requirements.txt
asgiref==3.6.0
astroid==2.12.13
attrs==22.2.0
black==22.12.0
certifi==2022.12.7
click==8.1.3
coverage==7.0.1
dill==0.3.6
distlib==0.3.6
Django==3.2.16
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-extensions==3.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.0
filelock==3.9.0
html5lib==1.1
importlib-metadata==5.2.0
iniconfig==1.1.1
isort==5.11.4
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
pipenv==2022.12.19
platformdirs==2.6.2
pluggy==1.0.0
pylint==2.15.9
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytest==7.2.0
python-decouple==3.6
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
typing-extensions==4.4.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
webencodings==0.5.1
wrapt==1.14.1
zipp==3.11.0
